Territory interfaces cleanup
