+++ beta-src/src/interfaces/index.ts
@@ -2,12 +2,9 @@ export * from "./Viewport";
 export * from "./CountryTableData";
 export * from "./TerritoryI";
 export * from "./map/Coordinates";
-export * from "./map/Dimensions";
 export * from "./map/TerritoryMapData";
-export * from "./map/BBox";
 export * from "./map/Label";
 export * from "./map/Texture";
-export * from "./map/UnitSlot";
 export * from "./state/MemberData";
 export * from "./map/UnitMeta";
 export * from "./state/UpdateOrder";
