--- beta-src/src/interfaces/map/BBox.ts
@@ -1,4 +1,0 @@
-import { Coordinates } from "./Coordinates";
-import { Dimensions } from "./Dimensions";
-
-export interface BBox extends Coordinates, Dimensions {}
