--- beta-src/src/interfaces/map/Dimensions.ts
@@ -1,4 +1,0 @@
-export interface Dimensions {
-  height: number;
-  width: number;
-}
