--- beta-src/src/interfaces/map/UnitSlot.ts
@@ -1,6 +1,0 @@
-import UnitSlotName from "../../types/map/UnitSlotName";
-import { Coordinates } from "./Coordinates";
-
-export interface UnitSlot extends Coordinates {
-  name: UnitSlotName;
-}
