--- beta-src/src/types/map/UnitSlotName.ts
@@ -1,7 +1,0 @@
-import GetArrayElementType from "../../utils/getArrayElementType";
-
-export const UnitSlotNames = ["main", "nc", "sc"] as const;
-
-type UnitSlotName = GetArrayElementType<typeof UnitSlotNames>;
-
-export default UnitSlotName;
